requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.5.0
librt==0.9.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mypy==1.20.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
paho-mqtt==2.1.0
pathspec==1.0.4
pillow==12.1.1
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
ruff==0.15.10
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
torch==2.11.0+cpu
torchvision==0.26.0+cpu
typing_extensions==4.15.0
ultralytics==8.4.36
ultralytics-thop==2.0.18
urllib3==2.6.3
